拆分 reasoning CLI 模块
